refactor(contentful): builders and factories moved
factories to create contentful & manage instances
with error reporting decorators
