+++ packages/botonic-plugin-contentful/src/cms/factories/index.ts
@@ -1,0 +1,1 @@
+export * from './content-factories'
